Rename db_config
