+++ data_warehouse/forklift/pipeline/flows/create_database.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from prefect import Flow, Parameter, case, task
 
-from forklift.db_config import create_datawarehouse_client
+from forklift.db_engines import create_datawarehouse_client
 from forklift.pipeline.shared_tasks.control_flow import check_flow_not_running
