+++ data_warehouse/forklift/pipeline/generic_tasks.py
@@ -10,7 +10,7 @@ from sqlalchemy.engine import Connection
 from sqlalchemy.sql import Select
 
 from forklift.config import SQL_SCRIPTS_LOCATION
-from forklift.db_config import create_datawarehouse_client, create_engine
+from forklift.db_engines import create_datawarehouse_client, create_engine
 from forklift.pipeline import utils
 from forklift.pipeline.processing import prepare_df_for_loading
 from forklift.pipeline.utils import get_table, psql_insert_copy
