+++ data_warehouse/forklift/pipeline/shared_tasks/infrastructure.py
@@ -2,7 +2,7 @@ import prefect
 from prefect import task
 from sqlalchemy import Table
 
-from forklift.db_config import create_engine
+from forklift.db_engines import create_engine
 from forklift.pipeline import utils
