+++ data_warehouse/tests/test_pipeline/test_flows/test_create_database.py
@@ -1,4 +1,4 @@
-from forklift.db_config import create_datawarehouse_client
+from forklift.db_engines import create_datawarehouse_client
 from forklift.pipeline.flows.create_database import flow
 from tests.mocks import mock_check_flow_not_running
